translator/requirements.txt
asgiref==3.7.2
cachetools==4.2.4
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==2.1.1
Django==4.2.4
django-binary-database-files==1.0.16
django-cors-headers==4.2.0
djangorestframework==3.14.0
dnspython==2.3.0
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.60.0
googletrans==3.1.0a0
grpcio==1.56.2
grpcio-status==1.48.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.8
numpy==1.25.2
pexpect==4.8.0
Pillow==9.2.0
protobuf==3.20.3
psycopg2==2.9.3
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pyarrow==12.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycryptodome==3.15.0
pylance==0.6.0
pytz==2018.9
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
Strip==0.0.1.dev14
typing_extensions==4.7.1
tzdata==2022.1
urllib3==1.24.3